connector: refacte routing logic
we try to remove edgePodCIDR and use podCIDR. then we don't use
aggregated route anymore. Instead we add a dummy interface
and add all edge rotues on it. They will be propagated into other
nodes. At the same time, the edge routes are added into table 220
with priority 220. It guides the traffic to ipsec tunnels.